+++ 3. Advanced Front-end Developer/02. 就业提升/15 ts基础/practice-pocker/package.json
@@ -3,7 +3,7 @@
     "@types/node": "^24.8.1"
   },
   "scripts": {
-    "build": "tsc",
+    "build": "rm -rf dist && tsc",
     "dev": "nodemon --watch src -e ts --exec \"ts-node\" src/index.ts"
   }
 }
