+++ 3. Advanced Front-end Developer/02. 就业提升/15 ts基础/practice-pocker/src/enums.ts
@@ -1,0 +1,21 @@
+export enum Color {
+  heart = "♥",
+  spade = "♠",
+  club = "♣",
+  diamond = "♦",
+}
+export enum Point {
+  A = "A",
+  two = "2",
+  three = "3",
+  four = "4",
+  five = "5",
+  six = "6",
+  seven = "7",
+  eight = "8",
+  nine = "9",
+  ten = "10",
+  J = "11",
+  Q = "12",
+  K = "13",
+}
