+++ 3. Advanced Front-end Developer/02. 就业提升/15 ts基础/practice-pocker/src/types.ts
@@ -1,0 +1,7 @@
+import { Color, Point } from "./enums";
+
+export type Deck = Card[];
+export type Card = {
+  color: Color;
+  point: Point;
+};
